+++ src/main/java/com/hand/hdsp/quality/infra/measure/field/MaxFixedValueMeasure.java
@@ -44,8 +44,8 @@ public class MaxFixedValueMeasure implements Measure {
         BatchResultRuleDTO batchResultRuleDTO = batchResultRuleDTOList.get(0);
         batchResultRuleDTO.setExpectedValue(batchPlanFieldLineDTO.getExpectedValue());
 
-        long actualValue = Long.parseLong(batchResultRuleDTO.getActualValue());
-        long expectedValue = Long.parseLong(batchPlanFieldLineDTO.getExpectedValue());
+        double actualValue = Long.parseLong(batchResultRuleDTO.getActualValue());
+        double expectedValue = Long.parseLong(batchPlanFieldLineDTO.getExpectedValue());
         MeasureUtil.fixedCompare(batchPlanFieldLineDTO.getCompareWay(), actualValue, expectedValue, warningLevelList, batchResultRuleDTO);
 
         return batchResultRuleDTO;
